GuiText now updatable and FPS now displayed in game window
The way ThinMatrix did his GuiText, you would have to create a new
GuiText every time you want to change the displayed text.  It's better
to update the values of an object than to create a new object, so, in
order to display the FPS onscreen, I moved the code that creates the
model for the text to GuiText and changed the TextMgr to use the new
functionality.  If GuiTexts create a new texture every time they are
updated, that will need to be fixed.