Fog changed. Collision detection now uses basic bounding box.
Trees use a bounding box that the player can stand on top of.  Not
realistic, just a test to see if the player can stand on a bounding box
whereas previously I used distance from point of origin as collision
detection which is spherical and only has so much use.